Aquilla x2 working mod addition
I'm using a kit so that I wouldn't have to assemble the big pulley for the z. This mod moves some things by 1.5mm so the belt alignment is likely off, and I recalibrated my z-axis.
Ran a dial indicator over it to check everything, it's been running very nicely.